find out all slots/intents combinations for Inform/Select
