+++ agents/demo2303/dialogue_aggregate/userInformedIntents.json
@@ -1,0 +1,14 @@
+{
+  "景点": {
+    "Turns": 1,
+    "Dialogues": 1
+  },
+  "酒店": {
+    "Turns": 1,
+    "Dialogues": 1
+  },
+  "餐馆": {
+    "Turns": 1,
+    "Dialogues": 1
+  }
+}
