+++ agents/demo2303/dialogue_aggregate/userInformedSlots.json
@@ -1,0 +1,14 @@
+{
+  "景点.评分": {
+    "Turns": 1,
+    "Dialogues": 1
+  },
+  "酒店.名称": {
+    "Turns": 1,
+    "Dialogues": 1
+  },
+  "餐馆.人均消费,餐馆.推荐菜": {
+    "Turns": 1,
+    "Dialogues": 1
+  }
+}
