+++ agents/demo2303/dialogue_aggregate/userSelectedSlots.json
@@ -1,0 +1,6 @@
+{
+  "景点.源领域": {
+    "Turns": 1,
+    "Dialogues": 1
+  }
+}
