+++ agents/test/dialogue_aggregate/userSelectedSlots.json
@@ -1,0 +1,14 @@
+{
+  "景点.源领域": {
+    "Turns": 79,
+    "Dialogues": 73
+  },
+  "酒店.源领域": {
+    "Turns": 51,
+    "Dialogues": 48
+  },
+  "餐馆.源领域": {
+    "Turns": 95,
+    "Dialogues": 79
+  }
+}
